Tests/LibWeb/Text/expected/SiteIsolation/cross-site-form-post-origin-and-referer.txt

Origin: <form-page-origin>
Referer: <form-page-origin>/
